# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API to manage anime seiyuus",
    "version": "1.0.0",
    "title": "API of Animes Seiyuus"
  },
  "host": "anime-seiyuus.onrender.com",
  "basePath": "/",
  "tags": [
    {
      "name": "animes",
      "description": "Anime Operations"
    },
    {
      "name": "users",
      "description": "User Operations"
    }
  ],
  "schemes": ["https"],
  "paths": {
    "/animes": {
      "get": {
        "tags": ["animes"],
        "summary": "Get all animes",
        "description": "Get all available animes",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Anime"
              }
            }
          },
          "500": {
            "description": "database error"
          }
        }
      },
      "post": {
        "tags": ["animes"],
        "summary": "Create new anime",
        "description": "Create new anime",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Anime object: must be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Anime"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "invalid title/year/chapters supplied"
          },
          "401": {
            "description": "notprivided/invalid token"
          },
          "500": {
            "description": "database error"
          }
        }
      }
    },
    "/animes/{id}": {
      "get": {
        "tags": ["animes"],
        "summary": "Find anime by ID",
        "description": "Returns a single anime",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of anime to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Anime"
            }
          },
          "400": {
            "description": "invalid ID supplied"
          },
          "404": {
            "description": "anime not found"
          },
          "500": {
            "description": "database error"
          }
        }
      },
      "put": {
        "tags": ["animes"],
        "summary": "Update an existing anime",
        "description": "Update data of an anime",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Anime that needs to be update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Anime"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Anime"
            }
          },
          "400": {
            "description": "invalid ID/title/year/chapters supplied"
          },
          "404": {
            "description": "anime not found"
          },
          "401": {
            "description": "notprivided/invalid token"
          },
          "500": {
            "description": "database error"
          }
        }
      },
      "delete": {
        "tags": ["animes"],
        "summary": "Deletes an anime",
        "description": "Deletes an anime by ID",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "invalid ID supplied"
          },
          "404": {
            "description": "anime not found"
          },
          "401": {
            "description": "not privided/invalid token"
          },
          "500": {
            "description": "database error"
          }
        }
      }
    },
    "/signup": {
      "post": {
        "tags": ["users"],
        "summary": "Register a new user",
        "description": "Creates a new user in the sistem",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Data user to sign up",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "invalid username/email/password supplied"
          },
          "500": {
            "description": "database error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["users"],
        "summary": "Logs user into the system",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Data to login",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "invalid email/password supplied"
          },
          "500": {
            "description": "database error"
          }
        }
      }
    }
  },
  "definitions": {
    "Anime": {
      "type": "object",
      "required": ["title", "year", "chapters"],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "year": {
          "type": "integer"
        },
        "chapters": {
          "type": "integer"
        }
      }
    },

    "User": {
      "type": "object",
      "required": ["username", "email", "password"],
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Login": {
      "type": "object",
      "required": ["username", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    }
  }
}